refactor: move to space domains
