+++ src/modules/space/domain/Space.ts
