+++ src/modules/space/repos/ISpaceRepository.ts
@@ -1,4 +1,4 @@
-import { DomainSpace } from '../../../graphql/schema/space/domain/Space'
+import { DomainSpace } from '../domain/Space'
 import { CreateSpaceDTO } from '../useCases/space/createSpace/CreateSpaceDTO'
 
 export interface ISpaceRepository {
